Modify PlayerInput and CropProperties
In preparation for sending the fruit to trucks,
PlayerInput.Harvest now sends the key that was
pressed.

Harvest sends a debug message when the crop is
harvested. This will eventually be a method that
updates the Truck fullness value, but for now, the
debug message returns the key that was used when
harvest progress was completed.

CropProperties.HarvestFruit now returns harvestProgress.