Add GetCorrectTruckID and FillTruck
UIManager.UpdateCorrectTruck calls
PlayerOverlay.UpdateCorrectTruck to update
truck progress.

The 4 interact keys - H,J,K, and L - correspond to
trucks 00, 01, 02, and 03.
PlayerOverlay.trucks is an array of the trucks. The index
corresponds to the truck ID.
GetCorrectTruckID takes the last interact key used to
harvest the fruit and returns the corresponding truck
index.

FillTruck uses the truck index to update the
corresponding truck gameobject. Currently, it only logs
a message to console.