Add methods to increment fruitBonusTracker
TruckProperties has an array that increments an
array element based on the fruit that has been
harvested.

Currently, this is a long chain where the id is returned by
PlayerMain calling PlayerInteract.CheckHarvestedFruitType(),
which in turn calls CropProperties.GetCurrentFruitType().
The fruitID is then passed through UIManager.UpdateTruck(),
which then passes on PlayerOverlay.LoadAndCheckTruck(),
which finally calls TruckProperties.IncrementBonusTracker().