Modify CropProperties, TruckProperties, PlayerOverlay, and CropManager
Classes have been modified to use
FruitPoolProperties to get data related
to fruit pool.

New methods were added to CropProperties
and TruckProperties so that
FruitPoolProperties instance is set by
the manager and so that each instance of
CropProperties and TruckProperties doesn't
have to find the FruitPoolProperties
instance via FindObjectOfType method.