Add TriggerEndgame method Modify Timer and PlayerOverlay class
UIManager now has TriggerEndgame method. At the
moment, it only logs a message to console.

Timer.Countdown now calls UIEndgame.TriggerEndgame.

Timer also has a new method, AddBonusTime, which
adds more time to timeLeft. The intention is for
time bonuses to be given when the percentage of
the fruit delivered by a single truck is of a
certain threshold.