Modified UIEndgame
The number of fruitInfo instances are stored in a list,
fruitInforList. An instance of fruitInfo is only created
when the total amount of delivered fruit exceeds
fruitInfoList.Count.

During the loop to set the fruitInfo text and sprites,
the number of fruitInfo instances to be used will
match the amount of fruit types delivered. Other
instances of fruitInfo will be disabled.